Filtrado de grados por sede

+++ CodeSchool/WebSchool/views.py
@@ -134,14 +134,27 @@ def delete_goal(request, id_Goal):
 
 @login_required(login_url='/')
 def grades(request):
-    grades_list = Grade.objects.all()
-    return render_to_response('grades.html',{'grades':grades_list}, context_instance=RequestContext(request))
+    if request.method == 'POST':
+        headquarters_list = Headquarter.objects.all()
+                
+        headquarter_selected = request.POST.getlist('headquarters')
+        headquarter = headquarter_selected[0]
+        
+        if headquarter == '-1':
+            grades_list = Grade.objects.all()
+        else:
+            grades_list = Grade.objects.filter(grade_headquarter = headquarter)        
+        
+        return render_to_response('grades.html',{'grades':grades_list, 'headquarters':headquarters_list, 'headquarter_selected':int(headquarter)}, context_instance=RequestContext(request))
+    else:
+        grades_list = Grade.objects.all()
+        headquarters_list = Headquarter.objects.all()
+    return render_to_response('grades.html',{'grades':grades_list, 'headquarters':headquarters_list, 'headquarter_selected':'-1'}, context_instance=RequestContext(request))
 
 @login_required(login_url='/')
 def add_grade(request):
     if request.method == 'POST':
-        form = GradeForm(request.POST)
-        
+        form = GradeForm(request.POST)        
         headquarter_selected = request.POST.getlist('grade_headquarter')
         headquarter = headquarter_selected[0]
         grades_select = request.POST.getlist('grade_name')
@@ -163,7 +176,14 @@ def edit_grade(request, id_grade):
     grade = Grade.objects.get(pk = id_grade)
     if request.method == 'POST':
         form = GradeForm(request.POST, instance = grade)
-        if form.is_valid():
+        headquarter_selected = request.POST.getlist('grade_headquarter')
+        headquarter = headquarter_selected[0]
+        grades_select = request.POST.getlist('grade_name')
+        grade = grades_select[0]        
+        sc = School.objects.get(pk = 1)
+        q = Grade.objects.filter(grade_year = sc.school_year).filter(grade_name = grade).filter(grade_headquarter = headquarter)
+        print(q)
+        if len(q)==0 and form.is_valid():
             form.save()
             return HttpResponseRedirect('/grades')
     else:
@@ -173,8 +193,10 @@ def edit_grade(request, id_grade):
 @login_required(login_url='/')
 def delete_grade(request, id_grade):
     grade = Grade.objects.get(pk = id_grade)
-    grade.delete()
-    return HttpResponseRedirect('/grades')
+    if request.method == 'POST':
+        grade.delete()
+        return HttpResponseRedirect('/grades')
+    return render_to_response('grades.html', {'grade':grade, 'delete':True}, context_instance = RequestContext(request))
 
 @login_required(login_url='/')
 def grades_history(request):
